requirements-agent.txt
# LiveKit Agent Requirements - Upgraded to v1.2.18 (latest stable)
livekit-agents==1.2.18

# LiveKit Plugins - Upgraded to v1.2.18 (latest stable)
livekit-plugins-deepgram==1.2.18
livekit-plugins-elevenlabs==1.2.18
livekit-plugins-openai==1.2.18
livekit-plugins-groq==1.2.18
livekit-plugins-silero==1.2.18
livekit-plugins-speechify==1.2.18
livekit-plugins-cartesia==1.2.18
livekit-plugins-turn-detector==1.2.18
mcp>=1.14.0,<2.0.0

# Voice/Audio Processing
pydub==0.25.1
sounddevice==0.5.1
numpy==1.26.4

# LLM/AI Libraries
openai==1.107.3
groq==0.14.0
deepgram-sdk==4.7.0
elevenlabs==2.8.0
tiktoken==0.9.0
anthropic==0.18.0

# Database and Storage
supabase==2.17.0
asyncpg==0.30.0
redis==5.2.0

# RAG/Embedding - Remote API services only
# Note: Using httpx for API calls (already included above)

# Remove local ML inference deps to avoid supervised_proc crashes

# HTTP and Async
httpx==0.27.2
aiohttp~=3.10
websockets>=12.0,<16
asyncio-mqtt==0.16.2
fastapi==0.110.1

# Utilities
python-dotenv==1.0.1
tenacity==9.1.2
colorlog==6.9.0
pydantic>=2.5.3

# Monitoring
prometheus-client>=0.22
structlog==23.2.0

# Utils
scipy==1.11.4

# Fix for numpy compatibility
numpy<2.0.0
